+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 ### Changed
+- Add Boilerplate code from [template](https://github.com/markonikoas/starter)
 - Update Readme
 - Added Changelog
 - Added Readme